# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.Login:
    description: Login request credentials
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  dto.LoginResponse:
    description: Login request response
    properties:
      token:
        type: string
      user_info:
        $ref: '#/definitions/dto.UserInfo'
    type: object
  dto.SignUp:
    description: SignUp request credentials
    properties:
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  dto.UserInfo:
    description: User info
    properties:
      create_at:
        example: "2023-04-01T15:04:05Z"
        type: string
      first_name:
        example: User first name
        type: string
      last_name:
        example: User last name
        type: string
      username:
        example: User's username'
        type: string
    type: object
  handler_errors.ErrorResponse:
    description: request error message
    properties:
      message:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: API Server for Goal tracker app
  title: Goal tracker
  version: "0.1"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Api for access to user privilege levels
      parameters:
      - description: login
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler_errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler_errors.ErrorResponse'
      summary: Sign in
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create an account in system
      parameters:
      - description: sign up
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SignUp'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler_errors.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler_errors.ErrorResponse'
      summary: Signup
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"